requirements.txt
albumentations==1.1.0
flatbuffers==1.12
sklearn==0.0
torch==1.11.0 # why does no-one pin their dependencies, 1.12 does not work
Pillow==9.0.0
